Refactor

+++ delete-collection.py
@@ -1,19 +1,28 @@
 import os
 from pymilvus import MilvusClient, client
 from MilvusUtils import MilvusUtils
+import sys
 
-collection_name = os.getenv("MY_COLLECTION_NAME") or "demo_collection"
-db_name = os.getenv("MY_DB_NAME")
-# print(f"{db_name}.db")
-# print(collection_name)
+if __name__ == "__main__":
+    if len(sys.argv) > 1:
+        collection_name = sys.argv[1] 
+    else:
+        collection_name = "demo_collection"
 
-client = MilvusClient( 
-    uri="http://localhost:19530",
-    token="root:Milvus"
-)
+    # db_name = os.getenv("MY_DB_NAME")
+    # print(f"{db_name}.db")
+    # print(collection_name)
 
-if client is not None:   
-    MilvusUtils.delete_collection(collection_name, client)
-    print(f"Collection: {collection_name} deleted successfully.")
-else:
-    print("Milvus client is not initialized.")
+    client = MilvusClient( 
+        uri="http://localhost:19530",
+        token="root:Milvus"
+    )
+
+    if client is not None: 
+        if client.has_collection(collection_name):
+            MilvusUtils.delete_collection(collection_name, client)
+            print(f"Collection: {collection_name} deleted successfully.")
+        else:
+            print(f"Collection: {collection_name} does not exist.")
+    else:
+        print("Milvus client is not initialized.")
